refactored architecture to three layered
