+++ SavannaApp.Business/Interfaces/IAnimalFactory.cs
@@ -1,6 +1,6 @@
 ﻿using SavannaApp.Data.Entities.Animals;
 
-namespace SavannaApp.Data.Interfaces
+namespace SavannaApp.Business.Interfaces
 {
     public interface IAnimalFactory
     {
