+++ SavannaApp.Business/Interfaces/IAnimalGroupManager.cs
@@ -1,6 +1,6 @@
-﻿using SavannaApp.Data.Interfaces.Game;
+﻿using SavannaApp.Data.Interfaces;
 
-namespace SavannaApp.Data.Interfaces
+namespace SavannaApp.Business.Interfaces
 {
     public interface IAnimalGroupManager
     {
