+++ SavannaApp.Business/Interfaces/IGameService.cs
@@ -1,4 +1,4 @@
-﻿namespace SavannaApp.Data.Interfaces
+﻿namespace SavannaApp.Business.Interfaces
 {
     public interface IGameService
     {
