+++ SavannaApp.Business/Services/AnimalCreationService.cs
@@ -1,9 +1,8 @@
-﻿using SavannaApp.Data.Entities.Animals;
+﻿using SavannaApp.Business.Interfaces;
+using SavannaApp.Data.Entities.Animals;
 using SavannaApp.Data.Interfaces;
-using SavannaApp.Data.Interfaces.Game;
-using SavannaApp.Data.Interfaces.Map;
 
-namespace SavannaApp.Data.Util
+namespace SavannaApp.Business.Services
 {
     public class AnimalCreationService(IAnimalFactory animalFactory, IMapManager mapManager) : IAnimalCreationService
     {
