+++ SavannaApp.Data/Interfaces/IMovement.cs
@@ -1,5 +1,4 @@
 ﻿using SavannaApp.Data.Entities.Animals;
-using SavannaApp.Data.Interfaces.Game;
 
 namespace SavannaApp.Data.Interfaces
 {
