+++ SavannaApp.Data/MovementStrategies/HunterMovement.cs
@@ -1,10 +1,8 @@
 ﻿using SavannaApp.Data.Constants;
 using SavannaApp.Data.Entities.Animals;
 using SavannaApp.Data.Interfaces;
-using SavannaApp.Data.Interfaces.Game;
-using SavannaApp.Data.Interfaces.Map;
 
-namespace SavannaApp.Data.Entities.MovementStrategies
+namespace SavannaApp.Data.MovementStrategies
 {
     public class HunterMovement(IMapManager mapManager) : IMovement
     {
@@ -44,6 +42,6 @@ namespace SavannaApp.Data.Entities.MovementStrategies
             return true;
         }
 
-        
+
     }
 }
