+++ SavannaApp.Tests/Entities/Animals/AnimalTests/Move.cs
@@ -1,10 +1,7 @@
-﻿using SavannaApp.Data.Entities;
-using SavannaApp.Data.Entities.Animals;
-using SavannaApp.Data.Entities.MovementStrategies;
+﻿using SavannaApp.Data.Entities.Animals;
+using SavannaApp.Data.Helpers.Map;
 using SavannaApp.Data.Interfaces;
-using SavannaApp.Data.Interfaces.Game;
-using SavannaApp.Data.Interfaces.Map;
-using SavannaApp.Data.Util;
+using SavannaApp.Data.MovementStrategies;
 
 namespace SavannaApp.Tests.Entities.Animals.AnimalTests
 {
