+++ SavannaApp.Tests/Entities/MapTests/IsPositionValid.cs
@@ -1,6 +1,4 @@
 ﻿using SavannaApp.Data.Entities.Animals;
-using SavannaApp.Data.Entities;
-using SavannaApp.Data.Interfaces.Game;
 using SavannaApp.Data.Interfaces;
 using Moq;
