+++ SavannaApp.Tests/Entities/MapTests/SetAnimal.cs
@@ -1,8 +1,6 @@
 ﻿using Moq;
-using SavannaApp.Data.Entities;
 using SavannaApp.Data.Entities.Animals;
 using SavannaApp.Data.Interfaces;
-using SavannaApp.Data.Interfaces.Game;
 
 namespace SavannaApp.Tests.Entities.MapTests
 {
