+++ SavannaApp.Tests/Entities/MovementStrategiesTests/Pray/Move.cs
@@ -1,11 +1,8 @@
 ﻿using SavannaApp.Data.Entities.Animals;
-using SavannaApp.Data.Entities.MovementStrategies;
-using SavannaApp.Data.Entities;
-using SavannaApp.Data.Interfaces.Game;
-using SavannaApp.Data.Interfaces.Map;
 using SavannaApp.Data.Interfaces;
-using SavannaApp.Data.Util;
 using Moq;
+using SavannaApp.Data.Helpers.Map;
+using SavannaApp.Data.MovementStrategies;
 
 namespace SavannaApp.Tests.Entities.MovementStrategiesTests.Pray
 {
