+++ SavannaApp.Tests/Entities/MovementStrategiesTests/Random/Move.cs
@@ -1,7 +1,6 @@
 ﻿using SavannaApp.Data.Entities.Animals;
-using SavannaApp.Data.Entities;
-using SavannaApp.Data.Interfaces.Game;
 using SavannaApp.Data.Interfaces;
+using SavannaApp.Data.MovementStrategies;
 
 namespace SavannaApp.Tests.Entities.MovementStrategiesTests.Random
 {
