+++ SavannaApp.Tests/SavannaApp.Tests.csproj
@@ -14,6 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\SavannaApp.Business\SavannaApp.Business.csproj" />
     <ProjectReference Include="..\SavannaApp.Data\SavannaApp.Data.csproj" />
   </ItemGroup>
