+++ SavannaApp.Tests/Util/AnimalCreationTests/CreateAnimal.cs
@@ -1,9 +1,8 @@
-﻿using SavannaApp.Data.Entities;
-using SavannaApp.Data.Util;
-using SavannaApp.Data.Entities.MovementStrategies;
-using SavannaApp.Data.Factory;
-using SavannaApp.Data.Interfaces.Game;
-using SavannaApp.Data.Entities.Animals;
+﻿using SavannaApp.Data.Entities.Animals;
+using SavannaApp.Data.Interfaces;
+using SavannaApp.Business.Services;
+using SavannaApp.Data.Helpers.Map;
+using SavannaApp.Data.MovementStrategies;
 
 namespace SavannaApp.Tests.Util.AnimalCreationTests
 {
