+++ SavannaApp.Tests/Util/MapManagerTests/GetBestFreeSpaceForPray.cs
@@ -1,9 +1,7 @@
 ﻿using Moq;
-using SavannaApp.Data.Interfaces.Game;
-using SavannaApp.Data.Interfaces.Map;
 using SavannaApp.Data.Interfaces;
-using SavannaApp.Data.Util;
 using SavannaApp.Data.Entities.Animals;
+using SavannaApp.Data.Helpers.Map;
 
 namespace SavannaApp.Tests.Util.MapManagerTests
 {
