+++ SavannaApp.Tests/Util/MapManagerTests/GetRandomFreePlaceOnMap.cs
@@ -1,9 +1,7 @@
 ﻿using Moq;
 using SavannaApp.Data.Entities.Animals;
-using SavannaApp.Data.Interfaces.Game;
-using SavannaApp.Data.Interfaces.Map;
+using SavannaApp.Data.Helpers.Map;
 using SavannaApp.Data.Interfaces;
-using SavannaApp.Data.Util;
 
 namespace SavannaApp.Tests.Util.MapManagerTests
 {
