+++ SavannaApp/UI/ConsoleInput.cs
@@ -1,6 +1,6 @@
-﻿using GameOfLife.Data.Interfaces.UI;
+﻿using SavannaApp.Business.Interfaces.UI;
 
-namespace GameOfLife.UI
+namespace SavannaApp.UI
 {
     public class ConsoleInput : IInputHandler
     {
